Fixes and optimizations

+++ src/data/planet_algorithms/algo0.rs
@@ -4,19 +4,19 @@ use super::PlanetAlgorithm;
 /// PlanetAlgorithm0 - All vertices at planet radius.
 /// This is the simplest algorithm.
 pub struct PlanetAlgorithm0 {
-    radius: f32,
+    radius: f64,
 }
 
 impl PlanetAlgorithm0 {
     pub fn new(planet: &Planet) -> Self {
         Self {
-            radius: planet.radius,
+            radius: planet.radius as f64,
         }
     }
 }
 
 impl PlanetAlgorithm for PlanetAlgorithm0 {
-    fn get_height(&self, _index: usize) -> f32 {
+    fn get_height(&self, _index: usize) -> f64 {
         self.radius
     }
 }
